refactor common_try
