+++ crates/sail-plan/src/extension/function/math/mod.rs
@@ -2,6 +2,7 @@ use std::cmp::{max, min};
 
 use datafusion::arrow::datatypes::{DataType, DECIMAL128_MAX_PRECISION, DECIMAL256_MAX_PRECISION};
 
+mod common_try;
 pub mod least_greatest;
 pub mod rand_poisson;
 pub mod randn;
